requirements.txt
asgiref==3.5.2
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
Django==4.1.4
djangorestframework==3.14.0
factory-boy==3.2.1
Faker==15.3.4
google-auth==2.15.0
idna==3.4
jwt==1.3.1
Pillow==9.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.13
